requirements.txt
streamlit>=1.30.0
psycopg2-binary>=2.9.9
requests>=2.31.0
pandas>=2.1.0
openpyxl>=3.1.0
